+++ example/test_app/urls.py
@@ -19,6 +19,7 @@ from . import views,auth
 urlpatterns = [
     path('admin/', admin.site.urls),
     path('passkeys/', include('passkeys.urls')),
+    path('api/passkeys/', include('passkeys.api.urls')),
     path('auth/login',auth.loginView,name="login"),
     path('auth/logout',auth.logoutView,name="logout"),